requirements.txt
Django==1.11.2
djangorestframework==3.6.3
pymongo==3.4.0
mongoengine